Refactor lincs_client into databases so tas can share access.
