+++ indra/sources/lincs_drug/processor.py
@@ -1,12 +1,11 @@
 from __future__ import absolute_import, print_function, unicode_literals
-from builtins import dict, str
 
 __all__ = ['LincsProcessor']
 
 import re
 
 from indra.statements import Agent, Inhibition, Evidence
-from .lincs_client import get_small_molecule_data, get_protein_data
+from indra.databases.lincs_client import get_small_molecule_data, get_protein_data
 
 
 class LincsProcessor(object):
